replace GetStakeActivation with CalStakeActivation

+++ task/era_update_active.go
@@ -57,7 +57,7 @@ func (task *Task) EraUpdataActive() error {
 	}
 
 	logrus.Infof("EraUpdateActive send tx hash: %s, stakeAccount: %s, stakeAccoutActive: %d, eraSnapshotActive: %d, eraProcessActive(old): %d, eraProcessActive(new): %d",
-		txHash, stakeAccount.ToBase58(), stakeAccountInfo.StakeAccount.Info.Stake.Delegation.Stake, eraActive, eraProcessActive, int64(eraProcessActive)+stakeAccountInfo.StakeAccount.Info.Stake.Delegation.Stake)
+		txHash, stakeAccount.ToBase58(), stakeAccountInfo.StakeAccount.Info.Stake.Delegation.Stake, eraActive, eraProcessActive, eraProcessActive+stakeAccountInfo.StakeAccount.Info.Stake.Delegation.Stake)
 
 	if err := task.waitTx(txHash); err != nil {
 		stakeManagerNew, err := task.client.GetStakeManager(context.Background(), task.cfg.StakeManagerAddress)
